refactor(STADTPULS-385): mark old types as deprecated
Preparation for generation of new types: https://supabase.io/docs/reference/javascript/generating-types
