+++ pages/sitemap.xml.ts
@@ -1,4 +1,4 @@
-import { ProjectsType } from "@common/types/supabase";
+import { ProjectsType } from "@common/types/supabase_DEPRECATED";
 import { getProjects } from "@lib/requests/getProjects";
 import { NextPage, NextApiResponse } from "next";
 import { Component } from "react";
