+++ src/components/CreateProjectForm/index.tsx
@@ -12,7 +12,7 @@ import {
   requiredProjectDescriptionValidation,
   requiredProjectConnectypeValidation,
 } from "@lib/formValidationUtil";
-import { ProjectsType } from "@common/types/supabase";
+import { ProjectsType } from "@common/types/supabase_DEPRECATED";
 
 type ChoosableProjectFieldsType = Pick<
   ProjectsType,
