+++ src/lib/requests/getProjectData/index.ts
@@ -1,5 +1,5 @@
 import { supabase } from "@auth/supabase";
-import { ProjectsType } from "@common/types/supabase";
+import { ProjectsType } from "@common/types/supabase_DEPRECATED";
 
 export const getProjectData = async (
   projectId: number
