+++ src/lib/requests/getProjects/index.ts
@@ -1,5 +1,5 @@
 import { supabase } from "@auth/supabase";
-import { ProjectsType } from "@common/types/supabase";
+import { ProjectsType } from "@common/types/supabase_DEPRECATED";
 
 export const getProjects = async (): Promise<ProjectsType[]> => {
   const { data: projects, error } = await supabase
