+++ src/mocks/supabaseData/index.ts
@@ -1,4 +1,4 @@
-import { DevicesType, ProjectsType } from "@common/types/supabase";
+import { DevicesType, ProjectsType } from "@common/types/supabase_DEPRECATED";
 import { fakeCategories } from "./categories";
 import { fakeProjects } from "./publicProjects";
 import { fakeDeviceRecords } from "./deviceRecords";
